Tests/LibWeb/Text/expected/HTML/SVGImageElement-load-and-error-events.txt

SUCCESS: "data:image/png;base64, iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO9TXL0Y4OHwAAAABJRU5ErkJggg=="
FAIL: "file:///i-do-not-exist"
FAIL: "https://something.invalid"
